+++ lib/lua.dart
@@ -1,3 +1,4 @@
 export 'src/ast.dart';
+export 'src/evaluator.dart' show evaluate;
 export 'src/formatter.dart' show format;
-export 'src/grammar.dart' show parse;
+// export 'src/grammar.dart' show parse;