+++ test/evaluate_test.dart
@@ -1,0 +1,16 @@
+import 'package:lua/src/ast.dart';
+import 'package:lua/src/evaluator.dart';
+import 'package:test/test.dart';
+
+void main() {
+  group('expressions', () {
+    test('basic literals', () {
+      expect(Nil().evaluate(), null);
+      expect(Bool(true).evaluate(), true);
+      expect(Bool(false).evaluate(), false);
+      expect(Number(1).evaluate(), 1);
+      expect(Number(1.2).evaluate(), 1.2);
+      expect(LuaString('x').evaluate(), 'x');
+    });
+  });
+}