+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1
 
 - Started working on `petitparser`-based parsing.
+- Started working on (AST) evaluator.
 
 ## 0.1.0